requirements.txt
fastapi
uvicorn
requests
python-dotenv
httpx
jinja2
starlette
itsdangerous
python-dotenv